dev: remove circular dependency
