+++ lib/dw/utils/getNonChartHeight.js
@@ -1,7 +1,14 @@
-import { outerHeight } from './index';
+export function outerHeight(element, withMargin = false) {
+    if (!element) return null;
+    let height = element.offsetHeight;
+    if (!withMargin) return height;
+    var style = getComputedStyle(element);
+    height += parseInt(style.marginTop) + parseInt(style.marginBottom);
+    return height;
+}
 /* global getComputedStyle */
 
-export default function() {
+export function getNonChartHeight() {
     let h = 0;
 
     const chart = document.querySelector('.dw-chart');
