+++ lib/dw/utils/index.js
@@ -6,7 +6,7 @@ import round from '@datawrapper/shared/round';
 import smartRound from '@datawrapper/shared/smartRound';
 import equalish from '@datawrapper/shared/equalish';
 import clone from '@datawrapper/shared/clone';
-import getNonChartHeight from './getNonChartHeight';
+import { outerHeight, getNonChartHeight } from './getNonChartHeight';
 import dateFormat from './dateFormat';
 import _ from 'underscore';
 
@@ -19,6 +19,7 @@ export {
     equalish,
     clone,
     getNonChartHeight,
+    outerHeight,
     dateFormat
 };
 
@@ -127,15 +128,6 @@ export function height(element) {
 
 export function width(element) {
     return parseFloat(getComputedStyle(element, null).width.replace('px', ''));
-}
-
-export function outerHeight(element, withMargin = false) {
-    if (!element) return null;
-    let height = element.offsetHeight;
-    if (!withMargin) return height;
-    var style = getComputedStyle(element);
-    height += parseInt(style.marginTop) + parseInt(style.marginBottom);
-    return height;
 }
 
 export function addClass(element, className) {
